+++ pandas/tests/io/formats/test_series_info.py
@@ -18,7 +18,7 @@ from pandas import (
 def test_info_categorical_column_just_works():
     n = 2500
     data = np.array(list("abcdefghij")).take(
-        np.random.default_rng(2).integers(0, 10, size=n)
+        np.random.default_rng(2).integers(0, 10, size=n, dtype=int)
     )
     s = Series(data).astype("category")
     s.isna()
